refactored api/agent/server.go
